V 0.0.2

+++ database/migrations/2020_07_12_013354_create_categorias_table.php
@@ -14,7 +14,7 @@ class CreateCategoriasTable extends Migration
     public function up()
     {
         Schema::create('categorias', function (Blueprint $table) {
-            $table->integer('id_categoria', true);
+            $table->bigInteger('id_categoria', true);
             $table->bigInteger('id_user')->nullable()->unsigned();
             $table->foreign('id_user')->references('id')->on('users')->onDelete('no action')->onUpdate('no action');
             $table->string('nom_categoria');
