+++ database/migrations/2020_07_13_180138_create_productos_table.php
@@ -14,14 +14,14 @@ class CreateProductosTable extends Migration
     public function up()
     {
         Schema::create('productos', function (Blueprint $table) {
-            $table->integer('id_producto',true)->unsigned();
-            $table->integer('id_proveedor')->nullable();
+            $table->bigInteger('id_producto',true)->unsigned();
+            $table->bigInteger('id_proveedor')->nullable();
             $table->bigInteger('id_user')->nullable()->unsigned();
             $table->foreign('id_user')->references('id')->on('users')->onDelete('no action')->onUpdate('no action');
             $table->foreign('id_proveedor')->references('id_proveedor')->on('proveedores')->onDelete('no action')->onUpdate('no action');
-            $table->integer('id_ubicacion')->nullable();
+            $table->bigInteger('id_ubicacion')->nullable();
             $table->foreign('id_ubicacion')->references('id_ubicacion')->on('ubicaciones')->onDelete('no action')->onUpdate('no action');
-            $table->integer('id_categoria')->nullable();
+            $table->BigInteger('id_categoria')->nullable();
             $table->foreign('id_categoria')->references('id_categoria')->on('categorias')->onDelete('no action')->onUpdate('no action');
             $table->string('codigo_de_barras', 80)->nullable();
             $table->string('ruta_imagen')->nullable()->default('default.png');
